feat: add getNewToken API function
