+++ src/components/Api/userApi.ts
@@ -21,6 +21,20 @@ export interface IUserResp {
   errMsg: string;
 }
 
+export interface ITokenResp {
+  message: string;
+  token: string;
+  refreshToken: string;
+  userId: string;
+  name?: string;
+}
+
+export interface ILoginResp {
+  isSuccess: boolean;
+  tokenResp: ITokenResp;
+  errMsg: string;
+}
+
 export const createUser = async (newUser: IUserLogin): Promise<IUserResp> => {
   const resp = await fetch(USERS_URL, {
     method: 'POST',
@@ -40,7 +54,7 @@ export const createUser = async (newUser: IUserLogin): Promise<IUserResp> => {
   };
 
   if (resp.status === OK) {
-    const data = await resp.json() as IUser;
+    const data = (await resp.json()) as IUser;
 
     userResp.isSuccess = true;
     Object.assign(userResp.user, data);
@@ -49,4 +63,36 @@ export const createUser = async (newUser: IUserLogin): Promise<IUserResp> => {
   }
 
   return userResp;
+};
+
+export const getNewToken = async (userId: string, refreshToken: string): Promise<ILoginResp> => {
+  const resp = await fetch(`${USERS_URL}/${userId}/tokens`, {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${refreshToken}`,
+      Accept: 'application/json',
+      'Content-Type': 'application/json',
+    },
+  });
+  const loginResp : ILoginResp = {
+    isSuccess: false,
+    tokenResp: {
+      message: '',
+      token: '',
+      refreshToken: '',
+      userId: '',
+    },
+    errMsg: '',
+  };
+
+  if (resp.status === OK) {
+    const data = (await resp.json()) as ITokenResp;
+
+    loginResp.isSuccess = true;
+    Object.assign(loginResp.tokenResp, data);
+  } else {
+    loginResp.errMsg = await resp.text();
+  }
+
+  return loginResp;
 };
